requirements.txt
Django
djangorestframework
django-filter
django-crispy-forms
crispy-bootstrap5